// Angular/shared/enums/operators.enum.ts
export enum OperatorsEnum {
    Equal = "eq",
    NotEqualTo = "nteq",
    GreaterThan = "gt",
    LessThan = "lt",
    GreaterThanOrEqualTo = "gteq",
    LessThanOrEqualTo = "lteq",
    In = "in",
    NotIn = "ntin",
    Contains = "contains",
    NotContains = "ntcontains",
    Between = "btw",
    To = "to"
}